Use `Integer` for indexes in `ClientBulkWriteException`.
JAVA-5527

+++ driver-core/src/main/com/mongodb/ClientBulkWriteException.java
@@ -44,7 +44,7 @@ public final class ClientBulkWriteException extends MongoServerException {
     @Nullable
     private final MongoException error;
     private final List<WriteConcernError> writeConcernErrors;
-    private final Map<Long, WriteError> writeErrors;
+    private final Map<Integer, WriteError> writeErrors;
     @Nullable
     private final ClientBulkWriteResult partialResult;
 
@@ -60,7 +60,7 @@ public final class ClientBulkWriteException extends MongoServerException {
     public ClientBulkWriteException(
             @Nullable final MongoException error,
             @Nullable final List<WriteConcernError> writeConcernErrors,
-            @Nullable final Map<Long, WriteError> writeErrors,
+            @Nullable final Map<Integer, WriteError> writeErrors,
             @Nullable final ClientBulkWriteResult partialResult,
             final ServerAddress serverAddress) {
         super(message(error, writeConcernErrors, writeErrors, partialResult, serverAddress), serverAddress);
@@ -79,7 +79,7 @@ public final class ClientBulkWriteException extends MongoServerException {
     private static String message(
             @Nullable final MongoException error,
             @Nullable final List<WriteConcernError> writeConcernErrors,
-            @Nullable final Map<Long, WriteError> writeErrors,
+            @Nullable final Map<Integer, WriteError> writeErrors,
             @Nullable final ClientBulkWriteResult partialResult,
             final ServerAddress serverAddress) {
         return "Client-level bulk write operation error on server " + serverAddress + "."
@@ -122,7 +122,7 @@ public final class ClientBulkWriteException extends MongoServerException {
      * @see ClientBulkWriteResult.Verbose#getUpdateResults()
      * @see ClientBulkWriteResult.Verbose#getDeleteResults()
      */
-    public Map<Long, WriteError> getWriteErrors() {
+    public Map<Integer, WriteError> getWriteErrors() {
         return writeErrors;
     }
